Break out functionality to internal class
